Started "Help" Section for building Sensors.
