+++ additional_files/BuildSensors.html
@@ -1,0 +1,138 @@
+<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
+<html>
+<head>
+	<meta http-equiv="content-type" content="text/html; charset=windows-1252"/>
+	<title></title>
+	<meta name="generator" content="LibreOffice 6.0.2.1 (Windows)"/>
+	<meta name="created" content="2018-09-19T09:31:31.834000000"/>
+	<meta name="changed" content="2018-09-19T12:20:54.703000000"/>
+	<style type="text/css">
+		@page { margin: 2cm }
+		p { margin-bottom: 0.25cm; line-height: 115% }
+		a:link { so-language: zxx }
+	</style>
+</head>
+<body lang="en-US" dir="ltr">
+<p align="center" style="margin-bottom: 0cm; line-height: 100%"><font size="4" style="font-size: 16pt"><u><b>Building
+a Sensor Unit to work with KootNet Sensors</b></u></font></p>
+<p align="center" style="margin-bottom: 0cm; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt"><i><b>** WIP: This will be
+updated with better instructions as time goes on **</b></i></font></p>
+<p align="center" style="margin-bottom: 0cm; line-height: 100%"><br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; line-height: 100%"><font size="4" style="font-size: 14pt"><u><b>Currently
+Supported Sensor&rsquo;s</b></u></font></p>
+<p align="center" style="margin-bottom: 0cm; line-height: 100%"><br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt"><a href="https://www.raspberrypi.org/products/">Raspberry
+Pi products</a>: Sense HAT</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt"><a href="https://shop.pimoroni.com/">Pimoroni
+products</a>: BH1745, BME680, Enviro pHAT, LSM303D, VL53L1X</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt"><i>Note: Additional Sensors
+can be added fairly easily, so long as they support the Raspberry Pi
+and Python3</i></font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-style: normal; line-height: 100%">
+<font size="4" style="font-size: 14pt"><u><b>Setting up a new Sensor</b></u></font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt">Step 1. Purchase a <a href="https://www.raspberrypi.org/products/">Raspberry
+Pi</a> (Get either a 3B+ or Zero W, depending on your needs)</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt">Step 2. Purchase desired
+Sensors from the supported list above (Or buy and program in your
+own)</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt">Step 3. Install your Sensor(s)
+on to the Raspberry Pi</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt">Step 4. <a href="https://www.raspberrypi.org/documentation/installation/installing-images/">Install
+Raspbian</a> on your Raspberry Pi</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt">Step 5.  Make sure you are
+connected to the Internet and run the following command in a terminal
+to download and start the installer</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt">Please note, this will
+automatically reboot the Raspberry Pi after completion</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt">wget
+<a href="http://dragonwarz.net/Stuffz/SensorUpgrades/upgrade/install_sensors.sh">http://dragonwarz.net/Stuffz/SensorUpgrades/upgrade/install_sensors.sh</a>
+&amp;&amp; sudo chmod +x install_sensors.sh &amp;&amp; sudo bash
+install_sensors.sh &amp;&amp; sudo reboot</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<font size="3" style="font-size: 12pt">After following the onscreen
+Instructions, your Sensor Unit should be ready to go!</font></p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+<p align="center" style="margin-bottom: 0cm; font-weight: normal; line-height: 100%; text-decoration: none">
+<br/>
+
+</p>
+</body>
+</html>
